Finished up User system. Might want to move login functionality to UserService instead of UserRestController
