+++ src/main/resources/application.properties
@@ -1,7 +1,16 @@
 spring.h2.console.enabled=true
+spring.datasource.password=pass
 spring.datasource.url=jdbc:mysql://127.0.0.1:3306/KinoXP
 spring.datasource.username=chatgpt
-spring.datasource.password=pass
+
+
+####Local testing stuff for Mikkel :PPP
+#spring.datasource.url=jdbc:mysql://127.0.0.1:3306/Kino_XP
+#spring.datasource.username=root
+#spring.datasource.password=root
+
+
+
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.show-sql=true
